pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "everyday-scripts"
version = "0.2.0"
description = "Sandip's everyday personal scripts"
authors = ["Sandip Bhattacharya <sandipb@showmethesource.org>"]
license = "Apache-2.0"
readme = "README.md"
include = ["CHANGELOG.md", "everyday_scripts/mime.types"]
classifiers = [
    'Development Status :: 3 - Alpha',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.9',
    "Topic :: Utilities",
]
repository = "https://github.com/sandipb/everyday-scripts"
keywords = []

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.1"
colorama = "^0.4.5"
PyYAML = "^6.0"
trio = "^0.21.0"
humanize = "^4.4.0"
pytz = "^2022.2.1"
prettytable = "^3.8.0"
google-cloud-bigquery = "^3.11.4"
click = "^8.1.7"
pygithub = "^1.59.1"
google-cloud-monitoring = "^2.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
mypy = "^0.971"
types-requests = "^2.28.11"
trio-typing = "^0.7.0"
types-PyYAML = "^6.0.11"
types-colorama = "^0.4.15"
types-pytz = "^2022.2.1.0"
ruff = "^0.1.4"
pyright = "^1.1.334"
pre-commit = "^3.5.0"
yamllint = "^1.32.0"

[tool.poetry.scripts]
split_manifest = 'everyday_scripts.split_manifest:main'
mmm = 'everyday_scripts.mmm:main'
filename_time_parse = 'everyday_scripts.filename_time_parse:main'
analyze_emails = 'everyday_scripts.analyze_emails:main'
manage_gcp_custom_metric = 'everyday_scripts.manage_gcp_custom_metrics:main'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 140
target-version = ['py311']

# https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 140
target-version = 'py311'

[tool.ruff.isort]
combine-as-imports = true
# ignore = ["E501"]
# [tool.ruff.lint.per-file-ignores]
# "__init__.py" = ["E402"]

[tool.pylint]
disable = ["missing-function-docstring", "missing-module-docstring"]

[tool.pylance]
pythonVersion = "3.11"
